+++ src/audio/pipeline/CMakeLists.txt
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof
+	pipeline-graph.c
+	pipeline-stream.c
+	pipeline-params.c
+	pipeline-xrun.c
+	pipeline-schedule.c
+)
